Api ii

+++ src/graphql/mutations/createRace.js
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 
 export const CREATE_RACE = gql`
-  mutation createRace($distance: Int!) {
+  mutation createRace($distance: Float!) {
     createRace(distance: $distance) {
       id
       distance
